custom_rgb_controller/requirements.txt
openrgb-python
PyQt6
numpy
qtawesome
soundcard
